+++ client/src/components/syntaxChecker/checker.ts
