+++ client/src/components/syntaxChecker/readme.md
